requirements.txt
Pillow==3.2.0
requests==2.10.0
ws4py==0.3.4
# pywin==0.3.1
# pypiwin32
pytz==2016.6.1